Move styles out of index.html.
Set up packages and loaders for Webpack to handle stylesheets.

Load up styles in the App.js root file.

Refactor signin component to accept props like rest of components. Validate props.

Change class name to get divs to reach edges of screen.

Display header only for signed in user. Otherwise, display empty div. shrug. Add more prop validation.

Import sign in button. Add splash div for non-signed-in users.

Add Less files and mixins and such.

Capitalize component because the word has earned it.

Add test for header for not-signed-in user. Block off tests for not-signed-in and signed-in user.

Add a location prop passed into the SignIn component to specify how to style it. Add elements to splash screen.

Add FB signin button styling.

Add and style that typical three high-level selling points on the front page.

Move the bars to the bottom of the page. Add bonkers styling to active and hover states for signin button. Shorten the features text to be on two lines, at least on my fly-ass display.

86 the stripes. Add some more spacing.

Add and style the Footer component for the front page.

Change ID name that was copied over from example.

Remove footer component from Home view since footer is now included at the Main level.

Fix CSS so footer is always at the bottom of the page. Include footer component in main, so it is available in all views.

Add Made In ATX note to footer.